replace regex path matches with path

+++ lib/Thruk/Controller/Root.pm
@@ -300,13 +300,19 @@ but if used not via fastcgi/apache, there is no way around
 
 =cut
 
-sub thruk_index : Regex('thruk$') {
+sub thruk_index : Path('/thruk/') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return if Thruk::Utils::choose_mobile($c, $c->stash->{'url_prefix'}."thruk/cgi-bin/mobile.cgi");
     if( scalar @{ $c->request->args } > 0 and $c->request->args->[0] ne 'index.html' ) {
         return $c->detach("default");
     }
+
+    # redirect from /thruk to /thruk/
+    if($c->request->path eq 'thruk') {
+        return $c->response->redirect($c->stash->{'url_prefix'}."thruk/");
+    }
+
     if( $c->stash->{'use_frames'} and !$c->stash->{'show_nav_button'} ) {
         return $c->detach("thruk_index_html");
     }
@@ -338,7 +344,7 @@ page: /thruk/index.html
 
 =cut
 
-sub thruk_index_html : Regex('thruk\/index\.html$') :MyAction('AddSafeDefaults') {
+sub thruk_index_html : Path('/thruk/index.html') :MyAction('AddSafeDefaults') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return if Thruk::Utils::choose_mobile($c, $c->stash->{'url_prefix'}."thruk/cgi-bin/mobile.cgi");
@@ -363,7 +369,7 @@ page: /thruk/side.html
 
 =cut
 
-sub thruk_side_html : Regex('thruk\/side\.html$') :MyAction('AddSafeDefaults') {
+sub thruk_side_html : Path('/thruk/side.html') :MyAction('AddSafeDefaults') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     Thruk::Utils::check_pid_file($c);
@@ -386,7 +392,7 @@ page: /thruk/frame.html
 
 =cut
 
-sub thruk_frame_html : Regex('thruk\/frame\.html') {
+sub thruk_frame_html : Path('/thruk/frame.html') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     # allowed links to be framed
@@ -433,7 +439,7 @@ page: /thruk/main.html
 
 =cut
 
-sub thruk_main_html : Regex('thruk\/main\.html$') {
+sub thruk_main_html : Path('/thruk/main.html') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
 
@@ -460,7 +466,7 @@ page: /thruk/changes.html
 
 =cut
 
-sub thruk_changes_html : Regex('thruk\/changes\.html') :MyAction('AddDefaults') {
+sub thruk_changes_html : Path('/thruk/changes.html') :MyAction('AddDefaults') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     $c->stash->{infoBoxTitle}            = 'Change Log';
@@ -481,7 +487,7 @@ page: /thruk/docs/
 
 =cut
 
-sub thruk_docs : Regex('thruk\/docs\/') :MyAction('AddDefaults') {
+sub thruk_docs : Path('/thruk/docs/') :MyAction('AddDefaults') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     if( scalar @{ $c->request->args } > 0 and $c->request->args->[0] ne 'index.html' ) {
@@ -507,7 +513,7 @@ page: /thruk/cgi-bin/tac.cgi
 
 =cut
 
-sub tac_cgi : Regex('thruk\/cgi\-bin\/tac\.cgi') {
+sub tac_cgi : Path('/thruk/cgi-bin/tac.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/tac/index');
@@ -521,7 +527,7 @@ page: /thruk/cgi-bin/status.cgi
 
 =cut
 
-sub status_cgi : Regex('thruk\/cgi\-bin\/status\.cgi') {
+sub status_cgi : Path('/thruk/cgi-bin/status.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/status/index');
@@ -535,7 +541,7 @@ page: /thruk/cgi-bin/cmd.cgi
 
 =cut
 
-sub cmd_cgi : Regex('thruk\/cgi\-bin\/cmd\.cgi') {
+sub cmd_cgi : Path('/thruk/cgi-bin/cmd.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/cmd/index');
@@ -549,7 +555,7 @@ page: /thruk/cgi-bin/outages.cgi
 
 =cut
 
-sub outages_cgi : Regex('thruk\/cgi\-bin\/outages\.cgi') {
+sub outages_cgi : Path('/thruk/cgi-bin/outages.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/outages/index');
@@ -563,7 +569,7 @@ page: /thruk/cgi-bin/avail.cgi
 
 =cut
 
-sub avail_cgi : Regex('thruk\/cgi\-bin\/avail\.cgi') {
+sub avail_cgi : Path('/thruk/cgi-bin/avail.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/avail/index');
@@ -577,7 +583,7 @@ page: /thruk/cgi-bin/trends.cgi
 
 =cut
 
-sub trends_cgi : Regex('thruk\/cgi\-bin\/trends\.cgi') {
+sub trends_cgi : Path('/thruk/cgi-bin/trends.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/trends/index');
@@ -591,7 +597,7 @@ page: /thruk/cgi-bin/history.cgi
 
 =cut
 
-sub history_cgi : Regex('thruk\/cgi\-bin\/history\.cgi') {
+sub history_cgi : Path('/thruk/cgi-bin/history.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/history/index');
@@ -605,7 +611,7 @@ page: /thruk/cgi-bin/summary.cgi
 
 =cut
 
-sub summary_cgi : Regex('thruk\/cgi\-bin\/summary\.cgi') {
+sub summary_cgi : Path('/thruk/cgi-bin/summary.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/summary/index');
@@ -619,7 +625,7 @@ page: /thruk/cgi-bin/histogram.cgi
 
 =cut
 
-sub histogram_cgi : Regex('thruk\/cgi\-bin\/histogram\.cgi') {
+sub histogram_cgi : Path('/thruk/cgi-bin/histogram.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/histogram/index');
@@ -633,7 +639,7 @@ page: /thruk/cgi-bin/notifications.cgi
 
 =cut
 
-sub notifications_cgi : Regex('thruk\/cgi\-bin\/notifications\.cgi') {
+sub notifications_cgi : Path('/thruk/cgi-bin/notifications.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/notifications/index');
@@ -647,7 +653,7 @@ page: /thruk/cgi-bin/showlog.cgi
 
 =cut
 
-sub showlog_cgi : Regex('thruk\/cgi\-bin\/showlog\.cgi') {
+sub showlog_cgi : Path('/thruk/cgi-bin/showlog.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/showlog/index');
@@ -661,7 +667,7 @@ page: /thruk/cgi-bin/extinfo.cgi
 
 =cut
 
-sub extinfo_cgi : Regex('thruk\/cgi\-bin\/extinfo\.cgi') {
+sub extinfo_cgi : Path('/thruk/cgi-bin/extinfo.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/extinfo/index');
@@ -675,7 +681,7 @@ page: /thruk/cgi-bin/config.cgi
 
 =cut
 
-sub config_cgi : Regex('thruk\/cgi\-bin\/config\.cgi') {
+sub config_cgi : Path('/thruk/cgi-bin/config.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/config/index');
@@ -689,7 +695,7 @@ page: /thruk/cgi-bin/job.cgi
 
 =cut
 
-sub job_cgi : Regex('thruk\/cgi\-bin\/job.cgi') :MyAction('AddSafeDefaults') {
+sub job_cgi : Path('/thruk/cgi-bin/job.cgi') :MyAction('AddSafeDefaults') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
 
@@ -704,7 +710,7 @@ page: /thruk/cgi-bin/test.cgi
 
 =cut
 
-sub test_cgi : Regex('thruk\/cgi\-bin\/test\.cgi') {
+sub test_cgi : Path('/thruk/cgi-bin/test.cgi') {
     my( $self, $c ) = @_;
     return if defined $c->{'canceled'};
     return $c->detach('/test/index');
